pyproject.toml
[project]
name = "dt-xml"
version = "0.1.0"
description = "Система поиска по таможенным декларациям ЕАЭС с использованием эмбедингов и реранкинга"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Aleksandr Mordvinov"}
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "lxml>=5.3.0",
    "xmltodict>=0.14.0",
    "sentence-transformers>=3.0.0",
    "qdrant-client>=1.12.0",
    "psycopg[binary]>=3.2.0",
    "sqlalchemy>=2.0.0",
    "langdetect>=1.0.9",
    "unidecode>=1.3.8",
    "pymorphy3>=1.2.0",
    "rank-bm25>=0.2.2",
    "numpy>=2.1.0",
    "python-multipart>=0.0.12",
    "pyyaml>=6.0.2",
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.6.0",
    "pyright>=1.1.400",
    "bandit>=1.7.10",
    "mypy>=1.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "build",
    "dist",
]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
